IOT-1583: Removing /W3 warning from resource/csdk/security.
I'm also refactoring some SConscript files in order to enable
/W3 and /WX only for specific directories and I'm adding
_CRT_NONSTDC_NO_WARNINGS to disable warnings about issues we cannot
change anyway, because the code has to build for multiple platforms.

Reviewed-on: https://gerrit.iotivity.org/gerrit/16329
Tested-by: jenkins-iotivity <jenkins@iotivity.org>

+++ resource/csdk/logger/include/logger.h
@@ -91,7 +91,7 @@ typedef enum {
  * @param[in]    buffer     pointer to buffer of bytes
  * @param[in]    bufferSize max number of byte in buffer
  */
-void OCLogBuffer(LogLevel level, const char * tag, const uint8_t * buffer, uint16_t bufferSize);
+void OCLogBuffer(LogLevel level, const char * tag, const uint8_t * buffer, size_t bufferSize);
 
 #define OCLog(level,tag,mes) LOG_(LOG_ID_MAIN, (level), (tag), mes)
 #define OCLogv(level,tag,fmt,args...) LOG_(LOG_ID_MAIN, (level),tag,fmt,##args)
@@ -147,7 +147,7 @@ void OCLogBuffer(LogLevel level, const char * tag, const uint8_t * buffer, uint1
      * @param buffer     - pointer to buffer of bytes
      * @param bufferSize - max number of byte in buffer
      */
-    void OCLogBuffer(LogLevel level, const char * tag, const uint8_t * buffer, uint16_t bufferSize);
+    void OCLogBuffer(LogLevel level, const char * tag, const uint8_t * buffer, size_t bufferSize);
 #else  // For arduino platforms
     /**
      * Initialize the serial logger for Arduino
